% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_median.R
\name{dere}
\alias{dere}
\title{Impute Missing Values with Median}
\usage{
dere(data)
}
\arguments{
\item{data}{A data frame containing numeric columns. Non-numeric columns will be ignored.}
}
\value{
A data frame with missing values in numeric columns replaced by the column median.
}
\description{
This function replaces missing values in numeric columns of a dataset with the median of each respective column.
}
\examples{
data <- data.frame(x = c(1, NA, 3), y = c(4, 5, NA))
dere(data)
}
